add initial map file and update Makefile; refactor movement functions

+++ get_next_line_utils.c
@@ -6,7 +6,7 @@
 /*   By: yel-qori <yel-qori@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/23 20:16:52 by yel-qori          #+#    #+#             */
-/*   Updated: 2024/11/24 14:59:33 by yel-qori         ###   ########.fr       */
+/*   Updated: 2025/10/09 11:27:19 by yel-qori         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -41,9 +41,9 @@ char	*ft_strdup(const char *s1)
 	return (dst);
 }
 
-size_t	ft_strlen(const char *str)
+int	ft_strlen(const char *str)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (str[i] != '\0')
